+++ boyiERPServer/src/main/java/com/boyi/controller/BaseSupplierController.java
@@ -5,6 +5,7 @@ import cn.hutool.core.map.MapUtil;
 import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.boyi.common.constant.DBConstant;
 import com.boyi.controller.base.BaseController;
 import com.boyi.controller.base.ResponseResult;
 import com.boyi.entity.*;
@@ -34,23 +35,6 @@ import java.util.Map;
 @RestController
 @RequestMapping("/baseData/supplier")
 public class BaseSupplierController extends BaseController {
-
-
-    /**
-     * 获取有效数据
-     */
-    @PostMapping("/getSearchAllValideData")
-    @PreAuthorize("hasAuthority('baseData:unit:list')")
-    public ResponseResult getSearchAllValideData() {
-        List<BaseSupplier> baseSuppliers = baseSupplierService.list(new QueryWrapper<BaseSupplier>().eq("status", 0));
-
-        ArrayList<Map<Object,Object>> returnList = new ArrayList<>();
-        baseSuppliers.forEach(obj ->{
-            Map<Object, Object> returnMap = MapUtil.builder().put("value",obj.getId()+" : "+obj.getName() ).put("id", obj.getId()).put("name", obj.getName()).map();
-            returnList.add(returnMap);
-        });
-        return ResponseResult.succ(returnList);
-    }
 
     /**
      * 获取全部
@@ -162,11 +146,32 @@ public class BaseSupplierController extends BaseController {
      */
     @PostMapping("/update")
     @PreAuthorize("hasAuthority('baseData:supplier:update')")
-    public ResponseResult update(Principal principal, @Validated @RequestBody BaseSupplier BaseSupplier) {
-        BaseSupplier.setUpdated(LocalDateTime.now());
-        BaseSupplier.setUpdateUser(principal.getName());
+    public ResponseResult update(Principal principal, @Validated @RequestBody BaseSupplier baseSupplier) {
+        baseSupplier.setUpdated(LocalDateTime.now());
+        baseSupplier.setUpdateUser(principal.getName());
         try {
-            baseSupplierService.updateById(BaseSupplier);
+
+            // 1. 查询以前的信息
+            BaseSupplier oldOne = baseSupplierService.getById(baseSupplier.getId());
+            if(!oldOne.getName().equals(baseSupplier.getName()) ||
+                    !oldOne.getMobile().equals(baseSupplier.getMobile())||
+                    !oldOne.getAddress().equals(baseSupplier.getAddress())){
+
+                // 2. 先查询是否有被价目表审核完成的引用，有则不能修改，
+                int count = baseSupplierMaterialService.count(new QueryWrapper<BaseSupplierMaterial>()
+                        .eq(DBConstant.TABLE_BASE_SUPPLIER_MATERIAL.SUPPLIER_ID_FIELDNAME, baseSupplier.getId())
+                                .eq(DBConstant.TABLE_BASE_SUPPLIER_MATERIAL.STATUS_FIELDNAME,DBConstant.TABLE_BASE_SUPPLIER_MATERIAL.STATUS_FIELDVALUE_0)
+                        );
+                if(count>0){
+                    log.info("供应商ID[{}]不能修改，存在{}个 审核完成的 采购价目记录",baseSupplier.getId(),count);
+                    return ResponseResult.fail("供应商ID["+baseSupplier.getId()+"]不能修改，存在"+count+"个 审核完成的 采购价目记录");
+                }
+
+                baseSupplierService.updateById(baseSupplier);
+                log.info("供应商ID[{}]更新成功，old{},new:{}.",baseSupplier.getId(),oldOne,baseSupplier);
+            }else {
+                return ResponseResult.fail("没有信息更改!");
+            }
             return ResponseResult.succ("编辑成功");
         } catch (DuplicateKeyException e) {
             log.error("供应商，更新异常",e);
@@ -184,6 +189,13 @@ public class BaseSupplierController extends BaseController {
 
         if(count > 0){
             return ResponseResult.fail("请先删除"+count+"条对应入库记录!");
+        }
+
+        int count2 = baseSupplierMaterialService.count(new QueryWrapper<BaseSupplierMaterial>()
+                .in("supplier_id", ids));
+
+        if(count2 > 0){
+            return ResponseResult.fail("请先删除"+count2+"条对应价目记录!");
         }
 
         baseSupplierService.removeByIds(Arrays.asList(ids));
